Change Docker File to use Pre-Made Pandas Image
In my previous docker image, there was a really time consuming step
when pandas was being install. It was problematic for a number of
reasons, to the point that I pulled it into its own step in the
container generation. I'm now switching to a premade file I found
which starts with python and pandas, avoiding this step. This also
allowed me to delete some other library imports that are now
baked from the start.