major refactor to support multiple tag tracking
To allow tracking of differing priority issues, I changed the api
access routines to all take the target label as an input. This also
required adding a routine to allow cyphering between the tag names
 as I understand then, and the tag ids that the google api uses and
requires for requests.

Once that was done, I hard-coded in the tags I'm interested in, and did
a minor refactor to the metrics analysis logic to support multitple
dataframes, with the index of each being hcanged to the label in
question.

Finally, I replaced the ave_squares metric with the sum_square metrics,
which I find more useful in its raw form.